+++ .gitignore
@@ -38,3 +38,4 @@ experiments/data/
 !experiments/data/00_get_datasets.ipynb
 .hypothesis
 experiments/results
+experiments/*.txt
