+++ experiments/config/saps/models.py
@@ -1,9 +1,8 @@
 import numpy as np
 from numpy import concatenate as cat
-from experiments.models.supercart import SuperCARTClassifier, SuperCARTRegressor
 from experiments.util import Model
-from imodels import GreedyTreeClassifier, GreedyTreeRegressor, C45TreeClassifier
-from imodels import RuleFitClassifier, RuleFitRegressor
+from imodels import GreedyTreeClassifier, GreedyTreeRegressor, C45TreeClassifier, SaplingSumClassifier
+from imodels import RuleFitClassifier, RuleFitRegressor, SaplingSumRegressor
 
 RULEFIT_DEFAULT_KWARGS_CLASSIFICATION = {'random_state': 0, 'max_rules': None, 'include_linear': False, 'alpha': 1}
 ESTIMATORS_CLASSIFICATION = (
@@ -11,7 +10,7 @@ ESTIMATORS_CLASSIFICATION = (
      for n in np.concatenate((np.arange(1, 19, 3), [25, 30]))],
     [Model('CART', GreedyTreeClassifier, 'max_depth', n)
      for n in np.arange(1, 6, 1)],
-    [Model('SuperCART', SuperCARTClassifier, 'max_rules', n)
+    [Model('SAPS', SaplingSumClassifier, 'max_rules', n)
      for n in cat((np.arange(1, 19, 3), [25, 30]))],
     [Model('Rulefit', RuleFitClassifier, 'n_estimators', n, other_params=RULEFIT_DEFAULT_KWARGS_CLASSIFICATION)
      for n in np.arange(1, 11, 1)],  # can also vary n_estimators and get a good spread
@@ -25,6 +24,6 @@ ESTIMATORS_REGRESSION = (
      for n in np.arange(1, 6, 1)],
     [Model('CART (MAE)', GreedyTreeRegressor, 'max_depth', n, other_params={'criterion': 'absolute_error'})
      for n in np.arange(1, 6, 1)],
-    [Model('SuperCART', SuperCARTRegressor, 'max_rules', n)
+    [Model('SAPS', SaplingSumRegressor, 'max_rules', n)
      for n in cat((np.arange(1, 19, 3), [25, 30]))],
 )
